Refactored java class file directories.
