--- code/app/src/main/java/com/example/eventgate/Admin/AdminActivity.java
@@ -1,16 +1,0 @@
-package com.example.eventgate.Admin;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-import com.example.eventgate.R;
-
-public class AdminActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_admin);
-    }
-}
